+++ src/main/java/commands/ExitCommand.java
@@ -1,6 +1,8 @@
 package commands;
 
+import access.Access;
 import manager.chapter.CardList;
+import storage.Storage;
 import ui.Ui;
 
 public class ExitCommand extends Command {
@@ -10,7 +12,7 @@ public class ExitCommand extends Command {
             + "Example: " + COMMAND_WORD + "\n";
 
     @Override
-    public void execute(CardList cards, Ui ui) {
+    public void execute(CardList cards, Ui ui, Access access, Storage storage) {
         ui.showExit();
     }
