+++ src/main/java/commands/HelpCommand.java
@@ -1,6 +1,8 @@
 package commands;
 
+import access.Access;
 import manager.chapter.CardList;
+import storage.Storage;
 import ui.Ui;
 
 public class HelpCommand extends Command {
@@ -10,7 +12,7 @@ public class HelpCommand extends Command {
             + "Example: " + COMMAND_WORD + "\n";
 
     @Override
-    public void execute(CardList cards, Ui ui) {
+    public void execute(CardList cards, Ui ui, Access access, Storage storage) {
         ui.showHelpList();
     }
