+++ src/main/java/commands/ListCommand.java
@@ -1,7 +1,9 @@
 package commands;
 
+import access.Access;
 import manager.card.Card;
 import manager.chapter.CardList;
+import storage.Storage;
 import ui.Ui;
 
 import java.util.ArrayList;
@@ -13,7 +15,7 @@ public class ListCommand extends Command {
             + "Example: " + COMMAND_WORD + "\n";
 
     @Override
-    public void execute(CardList cards, Ui ui) {
+    public void execute(CardList cards, Ui ui, Access access, Storage storage) {
         ArrayList<Card> allCards = cards.getAllCards();
         int cardCount = cards.getCardCount();
         ui.showCardList(allCards, cardCount);
