+++ src/main/java/commands/ReviseCommand.java
@@ -1,9 +1,11 @@
 package commands;
 
+import access.Access;
 import manager.card.Card;
 import manager.chapter.CardList;
 import manager.chapter.Chapter;
 import scheduler.Scheduler;
+import storage.Storage;
 import ui.Ui;
 
 import java.util.ArrayList;
@@ -36,7 +38,7 @@ public class ReviseCommand extends Command {
     }
 
     @Override
-    public void execute(CardList cards, Ui ui) {
+    public void execute(CardList cards, Ui ui, Access access, Storage storage) {
         ArrayList<Card> allCards = cards.getAllCards();
         ArrayList<Card> repeatCards = new ArrayList<>();
         int cardCount = cards.getCardCount();
