--- src/main/java/seedu/duke/command/Command.java
@@ -1,19 +1,0 @@
-package seedu.duke.command;
-
-import seedu.duke.tool.Access;
-import seedu.duke.tool.Storage;
-import seedu.duke.tool.Ui;
-
-public abstract class Command {
-    String fullCommand;
-
-    public Command(String fullCommand) {
-        this.fullCommand = fullCommand;
-    }
-
-    public abstract void execute(Access access, Ui ui, Storage storage);
-
-    public boolean isExit() {
-        return false;
-    };
-}
