+++ src/main/java/ui/Ui.java
@@ -1,5 +1,6 @@
 package ui;
 
+import access.Access;
 import commands.AddCommand;
 import commands.ExitCommand;
 import commands.ListCommand;
@@ -25,6 +26,7 @@ import static commands.ReviseCommand.MESSAGE_SHOW_RATING_PROMPT;
 public class Ui {
     private final Scanner in;
     private final PrintStream out;
+    private static final String lineSplit = "    ____________________________________________________________";
 
     public Ui() {
         this(System.in, System.out);
@@ -45,6 +47,10 @@ public class Ui {
 
     public void showWelcome() {
         out.println("Welcome to Kaji!\n");
+    }
+
+    public void showLevel(Access access) {
+        out.println(access.getLevel());
     }
 
     public void printEmptyLine() {
@@ -106,5 +112,9 @@ public class Ui {
         out.println("3. " + HelpCommand.MESSAGE_USAGE);
         out.println("4. " + AddCommand.MESSAGE_USAGE);
         out.println("5. " + ExitCommand.MESSAGE_USAGE);
+    }
+
+    public void showLine() {
+        System.out.println(lineSplit);
     }
 }
